requirements.txt
# Core pySigma and Elasticsearch dependencies - pinned for compatibility
pysigma==0.11.23
pysigma-backend-elasticsearch==1.1.6
elasticsearch==7.10.1

# Testing framework
pytest==7.4.4

# Code quality tools
black==24.10.0
flake8==7.1.1
isort==5.13.2

# JSON schema validation
jsonschema==4.23.0

# Compatible dependency versions to avoid conflicts
urllib3==1.26.18
jinja2==3.1.6
pyyaml==6.0.2
typing-extensions==4.12.2 